Fix arithmetic of MIDI number reassignment based on rewind hole position
Previously, there was a bug in the adjustment of each hole's MIDI number that
is done when realigning the tracker columns and MIDI numbers based on an
observed discrepancy between the detected rewind hole's MIDI number and the
MIDI number the rewind hole should have given the roll type. The code shifted
the MIDI numbers in the wrong direction, i.e., it interpreted the sign of the
value of `shifting` incorrectly. This value should be *subtracted* from the
MIDI number of each hole to correct the holes' alignment with the virtual
tracker bar, and conversely it should be *added* to the value of the mapping
from MIDI numbers to tracker bar columns (this part was already being done
correctly).