Mark Ampico A/B and Duo-Art options as enabled but experimental
The in-code caveats are also removed; the settings are no longer tentative because they've been tested and found to work for a handful of rolls of each type. The relatively high number of tracker column alignment failures noted in testing of each new roll type, however, indicates that the features should be labeled as experimental.