Reuduce expected margin before first tracker column for 88-note rolls
This causes all of the new gen2 88-note rolls to be aligned properly (without having to try to make the rewind hole detection work). Still needs testing to ensure it doesn't cause problems with the gen1 88-note rolls, or all of the other gen1 scans for that matter - because there has been one adjustment to the alignment logic for all roll types.